# apps/server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tabme API",
    "version": "1.0.0",
    "description": "The backend API for the Tabme Chrome Extension. It provides services for user authentication, data storage, and synchronization of spaces, folders, bookmarks, and sticky notes across devices."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address"
          },
          "subscriptionStatus": {
            "type": "string",
            "enum": [
              "free",
              "pro"
            ],
            "description": "User subscription status"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "User creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "User last update timestamp"
          }
        }
      },
...
      "Space": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Space ID"
          },
          "title": {
            "type": "string",
            "description": "Space title"
          },
          "position": {
            "type": "string",
            "description": "Position for ordering spaces"
          },
          "folders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Folder"
            }
          },
          "widgets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Widget"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Folder": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Folder ID"
          },
          "spaceId": {
            "type": "string",
            "format": "uuid",
            "description": "Space ID this folder belongs to"
          },
          "title": {
            "type": "string",
            "description": "Folder title"
          },
          "color": {
            "type": "string",
            "description": "Folder color"
          },
          "position": {
            "type": "string",
            "description": "Position for ordering folders"
          },
          "twoColumn": {
            "type": "boolean",
            "description": "Whether folder displays in two columns"
          },
          "collapsed": {
            "type": "boolean",
            "description": "Whether folder is collapsed"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookmarkItem"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BookmarkItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Bookmark item ID"
          },
          "folderId": {
            "type": "string",
            "format": "uuid",
            "description": "Folder ID this item belongs to"
          },
          "title": {
            "type": "string",
            "description": "Bookmark title"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Bookmark URL"
          },
          "favIconUrl": {
            "type": "string",
            "format": "uri",
            "description": "Favicon URL"
          },
          "position": {
            "type": "string",
            "description": "Position for ordering items"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Widget": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Widget ID"
          },
          "spaceId": {
            "type": "string",
            "format": "uuid",
            "description": "Space ID this widget belongs to"
          },
          "pos": {
            "type": "object",
            "properties": {
              "x": {
                "type": "number"
              },
              "y": {
                "type": "number"
              },
              "w": {
                "type": "number"
              },
              "h": {
                "type": "number"
              },
              "zIndex": {
                "type": "number"
              }
            }
          },
          "content": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "sticky-note"
                ]
              },
              "text": {
                "type": "string"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/folders/space/{spaceId}": {
      "get": {
        "tags": [
          "Folders"
        ],
        "summary": "Get folders in a space",
        "description": "Retrieve all folders within a specific space",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "spaceId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The space ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of folders in the space",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Folder"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Space not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/spaces": {
      "get": {
        "tags": [
          "Spaces"
        ],
        "summary": "Get user spaces",
        "description": "Retrieve all spaces for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user spaces",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Space"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "Spaces"
        ],
        "summary": "Create a new space",
        "description": "Create a new space for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "position"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "The space title"
                  },
                  "position": {
                    "type": "string",
                    "description": "The position for ordering spaces"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Space created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Space"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/spaces/{spaceId}": {
      "get": {
        "tags": [
          "Spaces"
        ],
        "summary": "Get space by ID",
        "description": "Retrieve a specific space by its ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "spaceId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The space ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Space details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Space"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Space not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "Spaces"
        ],
        "summary": "Update a space",
        "description": "Update an existing space",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "spaceId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The space ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "The space title"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Space updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Space"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Space not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Spaces"
        ],
        "summary": "Delete a space",
        "description": "Delete an existing space",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "spaceId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The space ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Space deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Space not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/spaces/positions": {
      "put": {
        "tags": [
          "Spaces"
        ],
        "summary": "Update space positions",
        "description": "Update the positions of multiple spaces for ordering",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "positions"
                ],
                "properties": {
                  "positions": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "spaceId",
                        "position"
                      ],
                      "properties": {
                        "spaceId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "position": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Positions updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/spaces/{spaceId}/users": {
      "get": {
        "tags": [
          "Spaces"
        ],
        "summary": "Get space users",
        "description": "Get all users that have access to a space",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "spaceId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The space ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of space users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Space not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "Spaces"
        ],
        "summary": "Add user to space",
        "description": "Add a user to a space with specified permissions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "spaceId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The space ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "targetUserId",
                  "position"
                ],
                "properties": {
                  "targetUserId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "The user ID to add to the space"
                  },
                  "position": {
                    "type": "string",
                    "description": "The position for the user in the space"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User added to space successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Space not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/spaces/{spaceId}/users/{targetUserId}": {
      "delete": {
        "tags": [
          "Spaces"
        ],
        "summary": "Remove user from space",
        "description": "Remove a user's access to a space",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "spaceId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The space ID"
          },
          {
            "in": "path",
            "name": "targetUserId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The user ID to remove from the space"
          }
        ],
        "responses": {
          "200": {
            "description": "User removed from space successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Space or user not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}